refactor: change test import
