+++ game/src/buildings/__tests__/bakery.test.ts
@@ -8,7 +8,7 @@ import {
   Tile,
 } from '../../types'
 import { initialState } from '../../state'
-import { bakery } from '../bakery'
+import { bakery } from '..'
 
 describe('buildings/bakery', () => {
   describe('bakery', () => {
