+++ game/src/buildings/__tests__/cloisterOffice.test.ts
@@ -9,7 +9,7 @@ import {
   Tableau,
   Tile,
 } from '../../types'
-import { cloisterOffice } from '../cloisterOffice'
+import { cloisterOffice } from '..'
 
 describe('buildings/cloisterOffice', () => {
   describe('use', () => {
