+++ game/src/buildings/__tests__/sacristy.test.ts
@@ -9,7 +9,7 @@ import {
   Tile,
 } from '../../types'
 
-import { sacristy } from '../sacristy'
+import { sacristy } from '..'
 
 describe('buildings/sacristy', () => {
   describe('use', () => {
