+++ game/src/buildings/__tests__/windmill.test.ts
@@ -8,7 +8,7 @@ import {
   Tableau,
   Tile,
 } from '../../types'
-import { windmill } from '../windmill'
+import { windmill } from '..'
 
 describe('buildings/windmill', () => {
   describe('windmill', () => {
